move ConstantTimeComparator to utils
